♻️  models.py 추출 — dataclass + intent ID 상수 (P0-L step 3)
변경 이유
- 6 모듈 분해의 leaf 모듈 부터 안전하게 추출.
  models 는 다른 5 모듈이 import 하는 base 이므로 가장 먼저 분리.

주요 변경 사항
- 새 파일 `engineering_conversation/models.py`:
  - 13 intent ID 상수 + `READ_ONLY_INTENTS` tuple
  - `EngineeringIntentMatch` / `EngineeringConversationResponse`
    dataclass
- `_legacy.py` 의 dataclass + 상수 블록 (53~131 줄, 약 85 줄)
  제거 → `from .models import ...` 로 re-export 만 남김.
- `__init__.py` facade 가 모델/상수 심볼을 `.models` 직접 import
  로 끌어옴 — `_legacy` 의존도를 줄여 다음 추출들이 더 작아짐.

비고
- 4657 PASS + 1 skipped 회귀 유지 (동작 변경 0).
- 다음 step (4): intent_detection.py 추출.
